+++ ipc/mex/quadrotor_msgs_BatteryVoltage.cc
@@ -1,6 +1,0 @@
-#include <ipc_bridge/msgs/quadrotor_msgs_BatteryVoltage.h>
-#include <ipc_bridge/matlab/quadrotor_msgs_BatteryVoltage.h>
-
-#define MESSAGE_NAME quadrotor_msgs::BatteryVoltage
-
-#include <ipc_bridge/mex_interface.h>