+++ ipc/ros/quadrotor_msgs_BatteryStatus_publisher.cc
@@ -1,11 +1,11 @@
 #include <ros/ros.h>
 #include <ipc_bridge/ipc_bridge.h>
 
-#include <quadrotor_msgs/BatteryVoltage.h>
-#include <ipc_bridge/msgs/quadrotor_msgs_BatteryVoltage.h>
+#include <quadrotor_msgs/BatteryStatus.h>
+#include <ipc_bridge/msgs/quadrotor_msgs_BatteryStatus.h>
 
 #define NAMESPACE quadrotor_msgs
-#define NAME BatteryVoltage
+#define NAME BatteryStatus
 
 ipc_bridge::Publisher<ipc_bridge::NAMESPACE::NAME> *p;
 ipc_bridge::NAMESPACE::NAME out_msg;